allowing nested Class/ActiveRecord compression
Before this patch DJ required me to pass in an ActiveRecord object
as a primary argument to the method call. This can make for
a really painful API because anything passed in an options hash, for
example, doesn't get the AR:Class compression and is stored in YAML.

This patch allows the compression even within a hash provided as an
argument.

Specs included.